+++ VNC.Domain/Enumerations/OrderStatusEnum.cs
@@ -10,7 +10,7 @@ namespace VNC.Domain.Enumerations
         public static readonly OrderStatusEnum Delivered = new(4, "Đã nhận hàng");
         public static readonly OrderStatusEnum Cancelled = new(5, "Đã hủy");
         protected OrderStatusEnum() { }
-        public OrderStatusEnum(int value, string name) : base(value, name)
+        public OrderStatusEnum(short value, string name) : base(value, name)
         {
         }
         public static IEnumerable<OrderStatusEnum> List() => new[] { Pending, Processing, Shipped, Delivered, Cancelled };
@@ -23,7 +23,7 @@ namespace VNC.Domain.Enumerations
             }
             return state;
         }
-        public static OrderStatusEnum FromValue(int value)
+        public static OrderStatusEnum FromValue(short value)
         {
             var state = List().SingleOrDefault(s => s.Value == value);
             if (state == null)
